[TASK] Bug Fix
